inserito nuovo endpoint da cui prendere loghi crypto

+++ server/api/utils.api.js
@@ -413,9 +413,13 @@ router.get("/fetchCryptoSvg", async (req, res, next) => {
             let symbolName = symbol.toLowerCase();
             lastSymbol = symbolName;
             try {
-                let result = await axios.get(`https://cryptoflash-icons-api.herokuapp.com/svg/icon/${symbolName}`);
-
-                let resp = fs.writeFileSync(`./symbolsSvg/${symbolName}.svg`, result.data);
+                //
+                // let result = await axios.get(`https://cryptoflash-icons-api.herokuapp.com/svg/icon/${symbolName}`);
+                let result = await axios.get(`https://logo.chainbit.xyz/${symbolName}`);
+                
+                if (!fs.existsSync(`./symbolsSvg/${symbolName}.svg`)) {
+                    let resp = fs.writeFileSync(`./symbolsSvg/${symbolName}.svg`, result.data);
+                  }
 
                 // console.log(resp);
                 // return res.json(result)
